// src/components/ProCharts/charting_library/charting_library.d.ts
export declare const widget: ChartingLibraryWidgetConstructor
export declare function version(): string
export declare type AccountId = Nominal<string, 'AccountID'>
export declare type ActionMetaInfo = ActionDescriptionWithCallback | MenuSeparator
export declare type AvailableSaveloadVersions = '1.0' | '1.1'
// ...
export declare type EmptyCallback = () => void
// ...
export declare type IBasicDataFeed = IDatafeedChartApi & IExternalDatafeed
export declare type INumberFormatter = IFormatter<number>
export declare type InputFieldValidator = (value: any) => InputFieldValidatorResult
export declare type InputFieldValidatorResult =
    | PositiveBaseInputFieldValidatorResult
    | NegativeBaseInputFieldValidatorResult
export declare type LanguageCode =
    | 'ar'
    | 'zh'
    | 'cs'
    | 'da_DK'
    | 'nl_NL'
    | 'en'
    | 'et_EE'
    | 'fr'
    | 'de'
    | 'el'
    | 'he_IL'
    | 'hu_HU'
    | 'id_ID'
    | 'it'
    | 'ja'
    | 'ko'
    | 'fa'
    | 'pl'
    | 'pt'
    | 'ro'
    | 'ru'
    | 'sk_SK'
    | 'es'
    | 'sv'
    | 'th'
    | 'tr'
    | 'vi'
// ...
export declare type PineJS = any
export declare type QuoteData = QuoteOkData | QuoteErrorData
export declare type QuotesCallback = (data: QuoteData[]) => void
export declare type ResolutionString = Nominal<string, 'ResolutionString'>
// ...
export declare type SupportedLineTools =
    | 'text'
    | 'anchored_text'
    | 'note'
    | 'anchored_note'
    | 'signpost'
    | 'double_curve'
    | 'arc'
    | 'icon'
    | 'arrow_up'
    | 'arrow_down'
    | 'arrow_left'
    | 'arrow_right'
    | 'price_label'
    | 'price_note'
    | 'arrow_marker'
    | 'flag'
    | 'vertical_line'
    | 'horizontal_line'
    | 'cross_line'
    | 'horizontal_ray'
    | 'trend_line'
    | 'info_line'
    | 'trend_angle'
    | 'arrow'
    | 'ray'
    | 'extended'
    | 'parallel_channel'
    | 'disjoint_angle'
    | 'flat_bottom'
    | 'pitchfork'
    | 'schiff_pitchfork_modified'
    | 'schiff_pitchfork'
    | 'balloon'
    | 'inside_pitchfork'
    | 'pitchfan'
    | 'gannbox'
    | 'gannbox_square'
    | 'gannbox_fixed'
    | 'gannbox_fan'
    | 'fib_retracement'
    | 'fib_trend_ext'
    | 'fib_speed_resist_fan'
    | 'fib_timezone'
    | 'fib_trend_time'
    | 'fib_circles'
    | 'fib_spiral'
    | 'fib_speed_resist_arcs'
    | 'fib_channel'
    | 'xabcd_pattern'
    | 'cypher_pattern'
    | 'abcd_pattern'
    | 'callout'
    | 'triangle_pattern'
    | '3divers_pattern'
    | 'head_and_shoulders'
    | 'fib_wedge'
    | 'elliott_impulse_wave'
    | 'elliott_triangle_wave'
    | 'elliott_triple_combo'
    | 'elliott_correction'
    | 'elliott_double_combo'
    | 'cyclic_lines'
    | 'time_cycles'
    | 'sine_line'
    | 'long_position'
    | 'short_position'
    | 'forecast'
    | 'date_range'
    | 'price_range'
    | 'date_and_price_range'
    | 'bars_pattern'
    | 'ghost_feed'
    | 'projection'
    | 'rectangle'
    | 'rotated_rectangle'
    | 'ellipse'
    | 'triangle'
    | 'polyline'
    | 'path'
    | 'curve'
    | 'cursor'
    | 'dot'
    | 'arrow_cursor'
    | 'eraser'
    | 'measure'
    | 'zoom'
    | 'brush'
    | 'highlighter'
// ...
export declare type ThemeName = 'Light' | 'Dark'
export declare type TimeFrameValue = TimeFramePeriodBack | TimeFrameTimeRange
export declare type Timezone = 'Etc/UTC' | CustomTimezones
// ...
export interface AccessList {
    type: 'black' | 'white'
    tools: AccessListItem[]
}
// ...
export interface ChartingLibraryWidgetConstructor {
    new (options: ChartingLibraryWidgetOptions | TradingTerminalWidgetOptions): IChartingLibraryWidget
}
export interface ChartingLibraryWidgetOptions {
    /** @deprecated */
    container_id?: string
    container: HTMLElement | string
    datafeed: IBasicDataFeed | (IBasicDataFeed & IDatafeedQuotesApi)
    interval: ResolutionString
    symbol?: string
    auto_save_delay?: number
    autosize?: boolean
    debug?: boolean
    disabled_features?: string[]
    drawings_access?: AccessList
    enabled_features?: string[]
    fullscreen?: boolean
    height?: number
    library_path?: string
    locale: LanguageCode
    numeric_formatting?: NumericFormattingParams
    saved_data?: object
    saved_data_meta_info?: SavedStateMetaInfo
    studies_access?: AccessList
    study_count_limit?: number
    symbol_search_request_delay?: number
    timeframe?: string
    timezone?: 'exchange' | Timezone
    toolbar_bg?: string
    width?: number
    charts_storage_url?: string
    charts_storage_api_version?: AvailableSaveloadVersions
    client_id?: string
    user_id?: string
    load_last_chart?: boolean
    studies_overrides?: StudyOverrides
    customFormatters?: CustomFormatters
    custom_formatters?: CustomFormatters
    overrides?: Overrides
    snapshot_url?: string
    preset?: 'mobile'
    time_frames?: TimeFrameItem[]
    custom_css_url?: string
    favorites?: Favorites
    save_load_adapter?: IExternalSaveLoadAdapter
    loading_screen?: LoadingScreenOptions
    settings_adapter?: ISettingsAdapter
    theme?: ThemeName
    compare_symbols?: CompareSymbol[]
    custom_indicators_getter?: (PineJS: PineJS) => Promise<readonly CustomIndicator[]>
}
export interface ClientSnapshotOptions {
    backgroundColor: string
    font: string
    fontSize: number
    legendMode: LegendMode
    hideResolution: boolean
}
export interface CompareSymbol {
    symbol: string
    title: string
}
export interface ContextMenuItem {
    position: 'top' | 'bottom'
    text: string
    click: EmptyCallback
}
export interface CreateButtonOptions {
    align: 'right' | 'left'
}
// ...
export interface CustomFormatters {
    timeFormatter: CustomFormatter
    dateFormatter: CustomFormatter
}
export interface CustomIndicator {
    readonly name: string
    readonly metainfo: any
    readonly constructor: any
}
// ...
export interface DialogParams<CallbackType> {
    title: string
    body: string
    callback: CallbackType
}
// ...
export interface Favorites {
    intervals: ResolutionString[]
    chartTypes: string[]
}
export interface FormatterParseResult {
    res: boolean
}
export interface GrayedObject {
    type: 'drawing' | 'study'
    name: string
}
// ...
export interface IChartingLibraryWidget {
    headerReady(): Promise<void>
    onChartReady(callback: EmptyCallback): void
    onGrayedObjectClicked(callback: (obj: GrayedObject) => void): void
    onShortcut(shortCut: string, callback: EmptyCallback): void
    subscribe<EventName extends keyof SubscribeEventsMap>(
        event: EventName,
        callback: SubscribeEventsMap[EventName]
    ): void
    unsubscribe<EventName extends keyof SubscribeEventsMap>(
        event: EventName,
        callback: SubscribeEventsMap[EventName]
    ): void
    chart(index?: number): IChartWidgetApi
    getLanguage(): LanguageCode
    setSymbol(symbol: string, interval: ResolutionString, callback: EmptyCallback): void
    remove(): void
    closePopupsAndDialogs(): void
    selectLineTool(linetool: SupportedLineTools): void
    selectedLineTool(): SupportedLineTools
    save(callback: (state: object) => void): void
    load(state: object, extendedData?: SavedStateMetaInfo): void
    getSavedCharts(callback: (chartRecords: SaveLoadChartRecord[]) => void): void
    loadChartFromServer(chartRecord: SaveLoadChartRecord): void
    saveChartToServer(onComplete?: EmptyCallback, onFail?: EmptyCallback, options?: SaveChartToServerOptions): void
    removeChartFromServer(chartId: string, onCompleteCallback: EmptyCallback): void
    onContextMenu(callback: (unixTime: number, price: number) => ContextMenuItem[]): void
    createButton(options?: CreateButtonOptions): HTMLElement
    showNoticeDialog(params: DialogParams<() => void>): void
    showConfirmDialog(params: DialogParams<(confirmed: boolean) => void>): void
    showLoadChartDialog(): void
    showSaveAsChartDialog(): void
    symbolInterval(): SymbolIntervalResult
    mainSeriesPriceFormatter(): INumberFormatter
    getIntervals(): string[]
    getStudiesList(): string[]
    addCustomCSSFile(url: string): void
    applyOverrides<TOverrides extends StudyOverrides>(overrides: TOverrides): void
    applyStudiesOverrides(overrides: object): void
    watchList(): IWatchListApi
    activeChart(): IChartWidgetApi
    chartsCount(): number
    layout(): LayoutType
    setLayout(layout: LayoutType): void
    layoutName(): string
    changeTheme(themeName: ThemeName, options?: ChangeThemeOptions): void
    getTheme(): ThemeName
    takeScreenshot(): void
    takeClientScreenshot(options?: Partial<ClientSnapshotOptions>): Promise<HTMLCanvasElement>
    lockAllDrawingTools(): IWatchedValue<boolean>
    hideAllDrawingTools(): IWatchedValue<boolean>
    magnetEnabled(): IWatchedValue<boolean>
    magnetMode(): IWatchedValue<number>
    symbolSync(): IWatchedValue<boolean>
    intervalSync(): IWatchedValue<boolean>
    crosshairSync(): IWatchedValue<boolean>
    timeSync(): IWatchedValue<boolean>
    startFullscreen(): void
    exitFullscreen(): void
    undoRedoState(): UndoRedoState
}
// ...
export interface IDatafeedQuotesApi {
    getQuotes(symbols: string[], onDataCallback: QuotesCallback, onErrorCallback: (msg: string) => void): void
    subscribeQuotes(
        symbols: string[],
        fastSymbols: string[],
        onRealtimeCallback: QuotesCallback,
        listenerGUID: string
    ): void
    unsubscribeQuotes(listenerGUID: string): void
}
// ...
export interface IExternalSaveLoadAdapter {
    getAllCharts(): Promise<ChartMetaInfo[]>
    removeChart<T extends number | string>(id: T): Promise<void>
    saveChart(chartData: ChartData): Promise<number>
    getChartContent(chartId: number): Promise<string>
    getAllStudyTemplates(): Promise<StudyTemplateMetaInfo[]>
    removeStudyTemplate(studyTemplateInfo: StudyTemplateMetaInfo): Promise<void>
    saveStudyTemplate(studyTemplateData: StudyTemplateData): Promise<void>
    getStudyTemplateContent(studyTemplateInfo: StudyTemplateMetaInfo): Promise<string>
    getDrawingTemplates(toolName: string): Promise<string[]>
    loadDrawingTemplate(toolName: string, templateName: string): Promise<string>
    removeDrawingTemplate(toolName: string, templateName: string): Promise<void>
    saveDrawingTemplate(toolName: string, templateName: string, content: string): Promise<void>
}
// ...
export interface ISettingsAdapter {
    initialSettings?: InitialSettingsMap
    setValue(key: string, value: string): void
    removeValue(key: string): void
}
// ...
export interface LoadingScreenOptions {
    foregroundColor?: string
    backgroundColor?: string
}
// ...
export interface NumericFormattingParams {
    decimal_sign: string
}
// ...
export interface Overrides {
    [key: string]: string | number | boolean
}
// ...
export interface SaveChartToServerOptions {
    chartName?: string
    defaultChartName?: string
}
export interface SaveLoadChartRecord {
    id: string
    name: string
    image_url: string
    modified_iso: number
    short_symbol: string
    interval: ResolutionString
}
export interface SavedStateMetaInfo {
    uid: number
    name: string
    description: string
}
// ...
export interface StudyOverrides {
    [key: string]: StudyOverrideValueType
}
export interface StudyTemplateData {
    name: string
    content: string
}
export interface StudyTemplateMetaInfo {
    name: string
}
export interface SubscribeEventsMap {
    toggle_sidebar: (isHidden: boolean) => void
    indicators_dialog: EmptyCallback
    toggle_header: (isHidden: boolean) => void
    edit_object_dialog: (params: EditObjectDialogEventParams) => void
    chart_load_requested: (savedData: object) => void
    chart_loaded: EmptyCallback
    mouse_down: (params: MouseEventParams) => void
    mouse_up: (params: MouseEventParams) => void
    drawing: (params: StudyOrDrawingAddedToChartEventParams) => void
    study: (params: StudyOrDrawingAddedToChartEventParams) => void
    undo: EmptyCallback
    redo: EmptyCallback
    undoRedoStackChanged: (state: UndoRedoState) => void
    reset_scales: EmptyCallback
    compare_add: EmptyCallback
    add_compare: EmptyCallback
    'load_study template': EmptyCallback
    onTick: (tick: Bar) => void
    onAutoSaveNeeded: EmptyCallback
    onScreenshotReady: (url: string) => void
    onMarkClick: (markId: Mark['id']) => void
    onPlusClick: (params: PlusClickParams) => void
    onTimescaleMarkClick: (markId: TimescaleMark['id']) => void
    onSelectedLineToolChanged: EmptyCallback
    layout_about_to_be_changed: (newLayoutType: LayoutType) => void
    layout_changed: EmptyCallback
    activeChartChanged: (chartIndex: number) => void
    drawing_event: (soursceId: string, drawingEventType: DrawingEventType) => void
    study_properties_changed: (id: EntityId) => void
    series_properties_changed: (id: EntityId) => void
    panes_height_changed: () => void
    panes_order_changed: () => void
}
// ...
export interface SymbolIntervalResult {
    symbol: string
    interval: ResolutionString
}
// ...
export interface TimeFrameItem {
    text: string
    resolution: ResolutionString
    description?: string
    title?: string
}